+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - Cookie support [#16](https://github.com/restbeast/restbeast/issues/16)
  - Apple Silicon M1 support for macOS
  - Enhance json decode error message
+ - new assertion function `assertNonEmptyString` [#35](https://github.com/restbeast/restbeast/issues/35)
 
 0.13.0
 - Request parameters [#13](https://github.com/restbeast/restbeast/issues/13)